move 'add_partition_date' function
